+++ draft/elf_loader/crates/elf/src/lib copy.rs
@@ -23,8 +23,8 @@ pub use program::Table as ProgramTable;
 pub use section::Header as SectionHeader;
 pub use section::Table as SectionTable;
 
-use x86_64::syscall;
 use x86_64::ToFlags;
+use x86_64::syscall;
 
 pub static mut IS_DYNAMIC: core::mem::MaybeUninit<bool> = core::mem::MaybeUninit::uninit();
 
@@ -85,7 +85,7 @@ impl ELF<'_> {
         loop {}
     }
 
-    pub unsafe fn read(file_descriptor: i32) -> Result<Self> {
+    pub unsafe fn read(file_descriptor: isize) -> Result<Self> {
         x86_64::print_str("\nReading Header\n");
         let header = ELFHeader::read(file_descriptor)?;
         x86_64::print_str("\nDone Reading Header\n");
